Fix header thumbnails failing on uploaded images
This was occasionally breaking because images in the left sidebar are
either uploaded or saved from an existing Solr image. Existing Solr
images have an id that can be translated to a URL to search Solr for
relevant ads, while uploaded images don't.

The long and short of it is, an image saved from Solr will show relevant
ads when rendered, and uploaded images won't.