Treat sidebar image models as uploaded or solr-backed
The notion here is given any image model, we should be able to tell if
it has a Solr counterpart (i.e. it actually came from saving an image
from a Solr backed search result), or if it was something the user
uploaded.

So imagespace.userData.images is a collection of ImageModel or
extensions of ImageModel.